package.json: add i18n scripts
Most notably i18n:lib to enable correct builds for activated internationalization.

The dependency to our webpack-i18n-tools is currently fixed to the latest version before the
switch to yarn 3 because installing it from a github repository fails otherwise. This version
suffices for the i18n:lib script. When it's installed from an actual in the future, this will
not be an issue anymore.